requirements.txt
Flask==2.3.3
spotipy==2.23.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Session==0.5.0
requests==2.31.0
python-jose==3.3.0 